Extract common game lobby function for Heist
