+++ HonorAmongThieves/Pages/Shared/_GameLobby.cshtml
@@ -1,0 +1,26 @@
+﻿<div id="pageName" class="display-4" style="text-decoration:underline; text-align:center">JOIN A GAME</div>
+<div class="state" id="startLobby" style="display:none">
+    <h3><u style="text-align:center">PLAYERS IN GAME</u></h3>
+    <div id="lobbyplayercount" class="h4"></div>
+    <ul id="lobbyList" class="h4"></ul>
+</div>
+<div class="row">&nbsp;</div>
+
+<div class="state" id="joinroom">
+    <div class="display-4" style="text-align:center">
+        Name: <input type="text" id="username" maxlength="10" size="10" oninput="this.value = this.value.toUpperCase()" /><br />
+        Room ID: <input type="text" id="roomid" maxlength="5" size="5" oninput="this.value = this.value.toUpperCase()" /><br /><br />
+        <input type="button" class="display-1" id="joinroombutton" value="JOIN" />
+    </div>
+    <h4>
+        <br />
+        Names are unique per room.<br />
+        Only letters and numbers allowed in username. Max 10 letters.<br />
+        You need a username to create a room.<br />
+        Room ID is automatically generated when you create a room. <br />
+    </h4>
+    <hr />
+    <div style="text-align:center" class="display-1">
+        <input type="button" id="createroombutton" value="CREATE" />
+    </div>
+</div>
