+++ HonorAmongThieves/Pages/Shared/_GameLobby.cshtml.cs
@@ -1,0 +1,16 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.RazorPages;
+
+namespace HonorAmongThieves.Pages.Shared
+{
+    public class _GameLobbyModel : PageModel
+    {
+        public void OnGet()
+        {
+        }
+    }
+}
