+++ HonorAmongThieves/wwwroot/js/gamelobby.js
@@ -1,0 +1,78 @@
+﻿var userName;
+var roomId;
+
+var reconnecting = true;
+
+connection.on("ShowError", function (errorMessage) {
+    window.alert(errorMessage);
+});
+
+// Create a new room in the lobby
+document.getElementById("createroombutton").addEventListener("click", function (event) {
+    var userNameIn = document.getElementById("username").value;
+
+    connection.invoke("CreateRoom", userNameIn).catch(function (err) {
+        return console.error(err.toString());
+    });
+    event.preventDefault();
+});
+
+// Update the page for the player when they join a room
+connection.on("JoinRoom_ChangeState", function (roomJoined, userJoined) {
+    // Hide the start button by default
+    var startButton = document.getElementById("startButtonDiv");
+    startButton.style.display = "none";
+
+    // This event is only sent to the caller on joining a room
+    userName = userJoined;
+    roomId = roomJoined;
+
+    // Write to session storage
+    sessionStorage.setItem("username", userName);
+    sessionStorage.setItem("roomid", roomId);
+
+    var elements = document.getElementsByClassName("state");
+    for (var i = 0; i < elements.length; i++) {
+        elements[i].style.display = "none";
+    }
+
+    document.getElementById("pageName").textContent = "LOBBY: " + roomJoined;
+    document.getElementById("startLobby").style.display = "block";
+});
+
+// Take over an existing user session
+connection.on("JoinRoom_TakeOverSession", function (roomJoined, userJoined) {
+    userName = userJoined;
+    roomId = roomJoined;
+})
+
+// Update the lobby state when a player joins
+connection.on("JoinRoom_UpdateState", function (playersConcat, userJoined) {
+    var playerList = document.getElementById("lobbyList");
+    playerList.innerHTML = "";
+    var players = playersConcat.split("|");
+    for (let i = 0; i < players.length; i++) {
+        var li = document.createElement("li");
+        li.textContent = players[i];
+        playerList.appendChild(li);
+    }
+
+    document.getElementById("lobbyplayercount").textContent = players.length + "/20";
+});
+
+// Joining a room in the lobby
+document.getElementById("joinroombutton").addEventListener("click", function (event) {
+    var userNameIn = document.getElementById("username").value;
+    var roomIdIn = document.getElementById("roomid").value;
+
+    connection.invoke("JoinRoom", roomIdIn, userNameIn).catch(function (err) {
+        return console.error(err.toString());
+    });
+    event.preventDefault();
+});
+
+// Game created - create the start button for the lobby owner
+connection.on("JoinRoom_CreateStartButton", function () {
+    var startButton = document.getElementById("startButtonDiv");
+    startButton.style.display = "block";
+});
